Take Register Command

+++ src/cogs/BasicCog.py
@@ -1,6 +1,10 @@
-from discord.ext import commands
 import asyncio
-import time
+import datetime
+
+from discord import File, channel
+from discord import utils
+from discord.ext import commands
+
 
 class BasicCog(commands.Cog):
 
@@ -57,7 +61,7 @@ class BasicCog(commands.Cog):
                          'sql', 'h', 'pdf', 'txt', 'cpp']
             filename = message.attachments[0].filename
             ext = filename.split(".")[-1]
-            if ext in Blacklist:
+            if ext in Blacklist and not isinstance(message.channel, channel.DMChannel):
                 await asyncio.sleep(1)
                 await message.delete()
                 await message.channel.send("Attachment Deleted. \n Please refrain from sending images or code on this "
@@ -76,6 +80,48 @@ class BasicCog(commands.Cog):
     # @commands.has_role("")
     async def clear(self, ctx, amount=5):
         await ctx.channel.purge(limit=amount)
+
+    @commands.command(name="Register", brief="Sends a list to the tutor of presents tutors",
+                      description="Gets a list of people who is present in the voice channel", aliases=["Reg","reg", "register", "pres", "present"])
+    @commands.cooldown(1, 2)
+    async def Register(self, ctx):
+        role = utils.find(lambda r: r.name == 'tutor' or r.name == 'Tutor', ctx.message.guild.roles)
+        members = []
+        tutor = []
+        try:
+            voice_channel = ctx.message.author.voice.channel
+            member_list = voice_channel.members
+            for i in member_list:
+                if role not in i.roles:
+                    members.append(i)
+                else:
+                    tutor.append(i)
+        except:
+            await ctx.send("You're not in a voice channel")
+
+        file_path = r"../src/csv/Register.txt"
+        f = open(file_path, "w")
+
+        date_time = datetime.datetime.now()
+        date = date_time.strftime("%d/%m/%Y, %H:%M:%S")
+        f.write(date)
+        f.write('\n')
+
+        f.write("Tutor: ")
+        sep = ""
+        for t in tutor:
+            f.write(t.name + sep)
+            sep = "&"
+        f.write("\n\n")
+
+        f.write("Present: \n")
+
+        for m in members:
+            f.write(m.name)
+            f.write("\n")
+        f.close()
+        file_name = ctx.guild.name + " " + date_time.strftime("%d/%m/%Y")
+        await ctx.author.send(file=File(file_path, filename=file_name))
 
 
 def setup(bot):
